realtime queue update
